+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 Changelog Application Jobsearch
 ===============================
+
 ## 1.6.1
 * Adds ML enriched locations to typeahead
 * Fixes a bug that wouldn't display number of ads for municipalities and regions in taxonomy search.
+* Adds header x-feature-include-synonyms-typeahead for choosing if enriched synonyms should be included in typeahead.
 
 ## 1.6.0
 * Introduces new lowercase type for request parsing